fixed nfields number in write 3d; in case of SGS_TKE these cant be written yet

+++ SAM6.10.10/SRC/SGS_EDMF/write_sgsfields3D.f90
@@ -6,7 +6,7 @@ use params
 
 implicit none
 
-real,intent(inout) :: nf
+integer,intent(inout) :: nf
 integer :: k,j,i
 character *80 long_name
 character *8 name
